+++ src/test/java/com/vaadin/ui/progressbar/tests/ProgressBarTest.java
@@ -15,11 +15,13 @@
  *
  */
 
-package com.vaadin.ui.progressbar;
+package com.vaadin.ui.progressbar.tests;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
+
+import com.vaadin.ui.progressbar.ProgressBar;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.*;
